fix: :bug: Correct DRF request parsing.
":" is used in places other than the "reading" signifier. Now we check
that the ":" is in the second position, signifying a "reading".

`str.replace` with two arguments will replace all instances of the first
argument with the second. In this case, we only want to replace the
first instance of the ":" with a "_" so we pass a third argument, which
limits the number of replacements to 1.